+++ Evergreen/Private/ConvertTo-DateTime.ps1
@@ -1,7 +1,7 @@
 Function ConvertTo-DateTime {
     <#
         .SYNOPSIS
-            Return string converted to date/time with formatting accounting for Windows PowerShell or PowerShell Core
+            Return a date/time string converted to a localised short date string.
     #>
     [OutputType([System.DateTime])]
     [CmdletBinding()]
@@ -11,23 +11,20 @@ Function ConvertTo-DateTime {
         [System.String] $DateTime,
 
         [Parameter(Position = 1)]
-        [System.String] $Pattern = 'MM/dd/yyyy HH:mm:ss'
+        [System.String] $Pattern = 'MM/dd/yyyy'
     )
 
-    # Return formatted DateTime if we're running on PowerShell Core vs. Windows PowerShell
-    # There's likely a better way to do this, but this is a start
+    # Convert the date/time passed to the function. If conversion fails, pass the same string back
     try {
-        If (Test-PSCore) {
-            $Output = [DateTime]::ParseExact($DateTime, $Pattern, [CultureInfo]::InvariantCulture)
-        }
-        Else {
-            $Output = [DateTime]::Parse($DateTime)
-        }
+        $ConvertedDateTime = [DateTime]::ParseExact($DateTime, $Pattern, [System.Globalization.CultureInfo]::CurrentUICulture.DateTimeFormat)
+        $Output = $ConvertedDateTime.ToShortDateString()
     }
     catch {
+        Write-Verbose -Message "$($MyInvocation.MyCommand): Failed to convert date, returning: [$DateTime]."
         $Output = $DateTime
     }
     
     # Write the output to the pipeline
+    Write-Verbose -Message "$($MyInvocation.MyCommand): Returning date: [$Output]."
     Write-Output -InputObject $Output
 }