+++ Evergreen/Private/Get-Architecture.ps1
@@ -15,7 +15,10 @@ Function Get-Architecture {
         "win64" { $architecture = "x64"; Break }
         "x86_64" { $architecture = "x64"; Break }
         "x64" { $architecture = "x64"; Break }
+        "64-bit" { $architecture = "x64"; Break }
+        "32-bit" { $architecture = "x86"; Break }
         "-x86" { $architecture = "x86"; Break }
+        "x86" { $architecture = "x86"; Break }
         "fxdependent" { $architecture = "fxdependent" }
         Default {
             Write-Verbose -Message "$($MyInvocation.MyCommand): Architecture not found, defaulting to x86."