+++ Evergreen/Public/Get-MicrosoftOffice.ps1
@@ -39,11 +39,9 @@ Function Get-MicrosoftOffice {
         $Json = $Content | ConvertFrom-Json
 
         # Build and array of the latest release and download URLs
-        # TODO: fix DateTime conversion
         $PSObject = [PSCustomObject] @{
             Version = $Json.AvailableBuild
-            #Date    = (ConvertTo-DateTime -DateTime $Json.TimestampUtc -Pattern $res.Get.DateTime)
-            Date    = $Json.TimestampUtc
+            Date    = ConvertTo-DateTime -DateTime $Json.TimestampUtc -Pattern $res.Get.DateTime
             Channel = $channel.Name
             URI     = $res.Get.DownloadUri
         }