+++ Evergreen/Public/Get-NotepadPlusPlus.ps1
@@ -28,52 +28,14 @@
     $res = Get-FunctionResource -AppName ("$($MyInvocation.MyCommand)".Split("-"))[1]
     Write-Verbose -Message $res.Name
 
-    # Get latest version and download latest Notepad++ release via GitHub API
-    # Query the Notepad++ repository for releases, keeping the latest stable release
+    # Get latest version and download latest release via GitHub API
     $iwcParams = @{
         Uri         = $res.Get.Uri
         ContentType = $res.Get.ContentType
-        Raw         = $True
     }
     $Content = Invoke-WebContent @iwcParams
 
-    If ($Null -ne $Content) {
-        $json = $Content | ConvertFrom-Json
-        $releases = $json | Where-Object { $_.prerelease -ne $True }
-        $latestRelease = $releases | Select-Object -First 1
-
-        # Build the output object with release details
-        ForEach ($release in $latestRelease.assets) {
-
-            # Filter for .exe
-            If (($release.content_type -eq "application/x-msdownload")) {
-
-                Switch -Regex ($release.browser_download_url) {
-                    "amd64" { $arch = "AMD64" }
-                    "arm64" { $arch = "ARM64" }
-                    "arm32" { $arch = "ARM32" }
-                    "x86_64" { $arch = "x86_64" }
-                    "x64" { $arch = "x64" }
-                    "-x86" { $arch = "x86" }
-                    Default { $arch = "x86" }
-                }
-
-                Switch -Regex ($release.browser_download_url) {
-                    "win" { $platform = "Windows" }
-                    Default { $platform = "Windows" }
-                }
-
-                # Build and array of the latest release and download URLs
-                $PSObject = [PSCustomObject] @{
-                    Version      = [RegEx]::Match($latestRelease.tag_name, $res.Get.MatchVersion).Captures.Groups[1].Value
-                    Platform     = $platform
-                    Architecture = $arch
-                    Date         = (ConvertTo-DateTime -DateTime $release.created_at)
-                    Size         = $release.size
-                    URI          = $release.browser_download_url
-                }
-                Write-Output -InputObject $PSObject
-            }
-        }
-    }
+    # Convert the returned release data into a useable object with Version, URI etc.
+    $object = ConvertFrom-GitHubReleasesJson -Content $Content -MatchVersion $res.Get.MatchVersion
+    Write-Output -InputObject $object
 }